requirements.txt
Django==1.11.13
pandas==0.20.1
requests==2.18.4
django_heroku==0.3.1
beautifulsoup4==4.6.0
lxml==4.2.1
Pygments==2.2.0
djangorestframework==3.8.2
gunicorn==19.8.1
dj-database-url==0.5.0
whitenoise==3.3.1
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==4.0.2
django-celery==4.1
django-jsonfield==1.0.1
kombu==3.0.37
pytz==2018.4